# typespec/tsp-output/@typespec/openapi3/openapi-1.0.0.yaml
openapi: 3.0.0
info:
  title: Julep API
  termsOfService: https://julep.ai/terms
  contact:
    name: Julep AI
    url: https://julep.ai
    email: developers@julep.ai
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  summary: A backend for creating stateful AI apps
  description: Julep is a backend for creating stateful AI apps with background tasks and long-term memory easily.
  version: 1.0.0
...
paths:
...
  /agents/{id}/docs:
    get:
      operationId: AgentDocsRoute_list
      description: List Docs owned by an Agent
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent
          schema:
            $ref: '#/components/schemas/Common.uuid'
        - $ref: '#/components/parameters/Common.PaginationOptions.limit'
        - $ref: '#/components/parameters/Common.PaginationOptions.offset'
        - $ref: '#/components/parameters/Common.PaginationOptions.sort_by'
        - $ref: '#/components/parameters/Common.PaginationOptions.direction'
        - $ref: '#/components/parameters/Common.PaginationOptions.metadata_filter'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Docs.Doc'
                required:
                  - items
    post:
      operationId: AgentDocsRoute_create
      description: Create a Doc for this Agent
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent resource
          schema:
            $ref: '#/components/schemas/Common.uuid'
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Common.ResourceCreatedResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Docs.CreateDocRequest'
...
  /agents/{id}/tasks/{child_id}:
    put:
      operationId: TasksRoute_update
      description: Update an existing task (overwrite existing values)
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent resource
          schema:
            $ref: '#/components/schemas/Common.uuid'
        - name: child_id
          in: path
          required: true
          description: ID of the resource to be updated
          schema:
            $ref: '#/components/schemas/Common.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Common.ResourceUpdatedResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tasks.UpdateTaskRequest'
    patch:
      operationId: TasksRoute_patch
      description: Update an existing task (merges with existing values)
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent resource
          schema:
            $ref: '#/components/schemas/Common.uuid'
        - name: child_id
          in: path
          required: true
          description: ID of the resource to be patched
          schema:
            $ref: '#/components/schemas/Common.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Common.ResourceUpdatedResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tasks.PatchTaskRequest'
    delete:
      operationId: TasksRoute_delete
      description: Delete a task by its id
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent resource
          schema:
            $ref: '#/components/schemas/Common.uuid'
        - name: child_id
          in: path
          required: true
          description: ID of the resource to be deleted
          schema:
            $ref: '#/components/schemas/Common.uuid'
      responses:
        '202':
          description: The request has been accepted for processing, but processing has not yet completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Common.ResourceDeletedResponse'
...
  /agents/{parent_id}/tasks/{id}:
    post:
      operationId: TasksCreateOrUpdateRoute_createOrUpdate
      description: Create or update a task
      parameters:
        - name: parent_id
          in: path
          required: true
          description: ID of the agent
          schema:
            $ref: '#/components/schemas/Common.uuid'
        - $ref: '#/components/parameters/Tasks.CreateOrUpdateTaskRequest.id'
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Common.ResourceUpdatedResponse'
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: '#/components/schemas/Tasks.CreateTaskRequest'
          text/x-yaml:
            schema:
              $ref: '#/components/schemas/Tasks.CreateTaskRequest'
          text/yaml:
            schema:
              $ref: '#/components/schemas/Tasks.CreateTaskRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/Tasks.CreateTaskRequest'
...
  /users/{id}/docs:
    get:
      operationId: UserDocsRoute_list
      description: List Docs owned by a User
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent
          schema:
            $ref: '#/components/schemas/Common.uuid'
        - $ref: '#/components/parameters/Common.PaginationOptions.limit'
        - $ref: '#/components/parameters/Common.PaginationOptions.offset'
        - $ref: '#/components/parameters/Common.PaginationOptions.sort_by'
        - $ref: '#/components/parameters/Common.PaginationOptions.direction'
        - $ref: '#/components/parameters/Common.PaginationOptions.metadata_filter'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Docs.Doc'
                required:
                  - items
    post:
      operationId: UserDocsRoute_create
      description: Create a Doc for this User
      parameters:
        - name: id
          in: path
          required: true
          description: ID of parent resource
          schema:
            $ref: '#/components/schemas/Common.uuid'
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Common.ResourceCreatedResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Docs.CreateDocRequest'
...
security:
  - ApiKeyAuth: []
  - ApiKeyAuth_: []
components:
  parameters:
...
    Common.PaginationOptions.direction:
      name: direction
      in: query
      required: true
      description: Sort direction
      schema:
        type: string
        enum:
          - asc
          - desc
        default: asc
      explode: false
    Common.PaginationOptions.limit:
      name: limit
      in: query
      required: true
      description: Limit the number of items returned
      schema:
        $ref: '#/components/schemas/Common.limit'
        default: 100
      explode: false
    Common.PaginationOptions.metadata_filter:
      name: metadata_filter
      in: query
      required: true
      description: Object to filter results by metadata
      schema:
        type: object
        additionalProperties:
          anyOf:
            - type: number
            - type: string
            - type: boolean
          nullable: true
      explode: false
    Common.PaginationOptions.offset:
      name: offset
      in: query
      required: true
      description: Offset the items returned
      schema:
        $ref: '#/components/schemas/Common.offset'
        default: 0
      explode: false
    Common.PaginationOptions.sort_by:
      name: sort_by
      in: query
      required: true
      description: Sort by a field
      schema:
        type: string
        enum:
          - created_at
          - updated_at
        default: created_at
      explode: false
...
    Tasks.CreateOrUpdateTaskRequest.id:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Common.uuid'
...
  schemas:
...
    Common.PyExpression:
      type: string
      description: A simple python expression compatible with SimpleEval.
    Common.ResourceCreatedResponse:
      type: object
      required:
        - id
        - created_at
        - jobs
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Common.uuid'
          description: ID of created resource
        created_at:
          type: string
          format: date-time
          description: When this resource was created as UTC date-time
          readOnly: true
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Common.uuid'
          description: IDs (if any) of jobs created as part of this request
          default: []
          readOnly: true
    Common.ResourceDeletedResponse:
      type: object
      required:
        - id
        - deleted_at
        - jobs
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Common.uuid'
          description: ID of deleted resource
        deleted_at:
          type: string
          format: date-time
          description: When this resource was deleted as UTC date-time
          readOnly: true
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Common.uuid'
          description: IDs (if any) of jobs created as part of this request
          default: []
          readOnly: true
    Common.ResourceUpdatedResponse:
      type: object
      required:
        - id
        - updated_at
        - jobs
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Common.uuid'
          description: ID of updated resource
        updated_at:
          type: string
          format: date-time
          description: When this resource was updated as UTC date-time
          readOnly: true
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Common.uuid'
          description: IDs (if any) of jobs created as part of this request
          default: []
          readOnly: true
...
    Common.limit:
      type: integer
      format: uint16
      minimum: 1
      maximum: 1000
      exclusiveMaximum: true
      description: Limit the number of results
...
    Common.offset:
      type: integer
      format: uint32
      minimum: 0
      description: Offset to apply to the results
    Common.uuid:
      type: string
      format: uuid
...
    Docs.CreateDocRequest:
      type: object
      required:
        - title
        - content
      properties:
        metadata:
          type: object
          additionalProperties: {}
        title:
          type: string
          maxLength: 800
          description: Title describing what this document contains
        content:
          anyOf:
            - type: string
            - type: array
              items:
                type: string
          description: Contents of the document
      description: Payload for creating a doc
    Docs.Doc:
      type: object
      required:
        - id
        - created_at
        - title
        - content
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Common.uuid'
          readOnly: true
        metadata:
          type: object
          additionalProperties: {}
        created_at:
          type: string
          format: date-time
          description: When this resource was created as UTC date-time
          readOnly: true
        title:
          type: string
          maxLength: 800
          description: Title describing what this document contains
        content:
          anyOf:
            - type: string
            - type: array
              items:
                type: string
          description: Contents of the document
        embeddings:
          anyOf:
            - type: array
              items:
                type: number
                format: float
            - type: array
              items:
                type: array
                items:
                  type: number
                  format: float
          description: Embeddings for the document
          readOnly: true
...
    Tasks.EvaluateStep:
      type: object
      required:
        - kind_
        - evaluate
      properties:
        kind_:
          type: string
          enum:
            - evaluate
          default: evaluate
          readOnly: true
        evaluate:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Common.PyExpression'
          description: The expression to evaluate
      allOf:
        - type: object
          required:
            - kind_
          properties:
            kind_:
              type: string
              enum:
                - evaluate
              description: The kind of step
              readOnly: true
          discriminator:
            propertyName: kind_
            mapping: {}
...
    Tasks.GetStep:
      type: object
      required:
        - kind_
        - get
      properties:
        kind_:
          type: string
          enum:
            - get
          default: get
          readOnly: true
        get:
          type: string
          description: The key to get
      allOf:
        - type: object
          required:
            - kind_
          properties:
            kind_:
              type: string
              enum:
                - get
              description: The kind of step
              readOnly: true
          discriminator:
            propertyName: kind_
            mapping: {}
...
    Tasks.PatchTaskRequest:
      type: object
      properties:
        description:
          type: string
          default: ''
        main:
          type: array
          items:
            anyOf:
              - $ref: '#/components/schemas/Tasks.EvaluateStep'
              - $ref: '#/components/schemas/Tasks.ToolCallStep'
              - $ref: '#/components/schemas/Tasks.PromptStepUpdateItem'
              - $ref: '#/components/schemas/Tasks.GetStep'
              - $ref: '#/components/schemas/Tasks.SetStep'
              - $ref: '#/components/schemas/Tasks.LogStep'
              - $ref: '#/components/schemas/Tasks.YieldStep'
              - $ref: '#/components/schemas/Tasks.ReturnStep'
              - $ref: '#/components/schemas/Tasks.SleepStep'
              - $ref: '#/components/schemas/Tasks.ErrorWorkflowStep'
              - $ref: '#/components/schemas/Tasks.YieldStep'
              - $ref: '#/components/schemas/Tasks.WaitForInputStep'
              - $ref: '#/components/schemas/Tasks.IfElseWorkflowStepUpdateItem'
              - $ref: '#/components/schemas/Tasks.SwitchStepUpdateItem'
              - $ref: '#/components/schemas/Tasks.ForeachStepUpdateItem'
              - $ref: '#/components/schemas/Tasks.ParallelStepUpdateItem'
              - type: object
                required:
                  - over
                  - map
                properties:
                  over:
                    allOf:
                      - $ref: '#/components/schemas/Common.PyExpression'
                    description: The variable to iterate over
                  map:
                    anyOf:
                      - $ref: '#/components/schemas/Tasks.EvaluateStep'
                      - $ref: '#/components/schemas/Tasks.ToolCallStep'
                      - $ref: '#/components/schemas/Tasks.PromptStepUpdateItem'
                      - $ref: '#/components/schemas/Tasks.GetStep'
                      - $ref: '#/components/schemas/Tasks.SetStep'
                      - $ref: '#/components/schemas/Tasks.LogStep'
                      - $ref: '#/components/schemas/Tasks.YieldStep'
                    description: The steps to run for each iteration
                  reduce:
                    allOf:
                      - $ref: '#/components/schemas/Common.PyExpression'
                    description: |-
                      The expression to reduce the results.
                      If not provided, the results are collected and returned as a list.
                      A special parameter named `results` is the accumulator and `_` is the current value.
                  initial:
                    description: The initial value of the reduce expression
                    default: []
                  parallelism:
                    type: integer
                    format: uint16
                    minimum: 1
                    maximum: 100
                    description: Whether to run the reduce expression in parallel and how many items to run in each batch
                allOf:
                  - type: object
                    properties:
                      kind_:
                        type: string
                        description: Discriminator property for BaseWorkflowStep.
                    discriminator:
                      propertyName: kind_
          minItems: 1
          description: The entrypoint of the task.
        input_schema:
          type: object
          additionalProperties: {}
          nullable: true
          description: The schema for the input to the task. `null` means all inputs are valid.
          default: null
        tools:
          type: array
          items:
            $ref: '#/components/schemas/Tasks.TaskTool'
          description: Tools defined specifically for this task not included in the Agent itself.
          default: []
        inherit_tools:
          type: boolean
          description: Whether to inherit tools from the parent agent or not. Defaults to true.
          default: true
        metadata:
          type: object
          additionalProperties: {}
      additionalProperties:
        type: array
        items:
          anyOf:
            - $ref: '#/components/schemas/Tasks.EvaluateStep'
            - $ref: '#/components/schemas/Tasks.ToolCallStep'
            - $ref: '#/components/schemas/Tasks.PromptStepUpdateItem'
            - $ref: '#/components/schemas/Tasks.GetStep'
            - $ref: '#/components/schemas/Tasks.SetStep'
            - $ref: '#/components/schemas/Tasks.LogStep'
            - $ref: '#/components/schemas/Tasks.YieldStep'
            - $ref: '#/components/schemas/Tasks.ReturnStep'
            - $ref: '#/components/schemas/Tasks.SleepStep'
            - $ref: '#/components/schemas/Tasks.ErrorWorkflowStep'
            - $ref: '#/components/schemas/Tasks.YieldStep'
            - $ref: '#/components/schemas/Tasks.WaitForInputStep'
            - $ref: '#/components/schemas/Tasks.IfElseWorkflowStepUpdateItem'
            - $ref: '#/components/schemas/Tasks.SwitchStepUpdateItem'
            - $ref: '#/components/schemas/Tasks.ForeachStepUpdateItem'
            - $ref: '#/components/schemas/Tasks.ParallelStepUpdateItem'
            - type: object
              required:
                - over
                - map
              properties:
                over:
                  allOf:
                    - $ref: '#/components/schemas/Common.PyExpression'
                  description: The variable to iterate over
                map:
                  anyOf:
                    - $ref: '#/components/schemas/Tasks.EvaluateStep'
                    - $ref: '#/components/schemas/Tasks.ToolCallStep'
                    - $ref: '#/components/schemas/Tasks.PromptStepUpdateItem'
                    - $ref: '#/components/schemas/Tasks.GetStep'
                    - $ref: '#/components/schemas/Tasks.SetStep'
                    - $ref: '#/components/schemas/Tasks.LogStep'
                    - $ref: '#/components/schemas/Tasks.YieldStep'
                  description: The steps to run for each iteration
                reduce:
                  allOf:
                    - $ref: '#/components/schemas/Common.PyExpression'
                  description: |-
                    The expression to reduce the results.
                    If not provided, the results are collected and returned as a list.
                    A special parameter named `results` is the accumulator and `_` is the current value.
                initial:
                  description: The initial value of the reduce expression
                  default: []
                parallelism:
                  type: integer
                  format: uint16
                  minimum: 1
                  maximum: 100
                  description: Whether to run the reduce expression in parallel and how many items to run in each batch
              allOf:
                - type: object
                  properties:
                    kind_:
                      type: string
                      description: Discriminator property for BaseWorkflowStep.
                  discriminator:
                    propertyName: kind_
      description: Payload for patching a task
...
    Tasks.TaskTool:
      type: object
      properties:
        inherited:
          type: boolean
          description: 'Read-only: Whether the tool was inherited or not. Only applies within tasks.'
          default: false
          readOnly: true
      allOf:
        - $ref: '#/components/schemas/Tools.CreateToolRequest'
...
    Tasks.UpdateTaskRequest:
      type: object
      required:
        - description
        - main
        - input_schema
        - tools
        - inherit_tools
      properties:
        description:
          type: string
          default: ''
        main:
          type: array
          items:
            anyOf:
              - $ref: '#/components/schemas/Tasks.EvaluateStep'
              - $ref: '#/components/schemas/Tasks.ToolCallStep'
              - $ref: '#/components/schemas/Tasks.PromptStep'
              - $ref: '#/components/schemas/Tasks.GetStep'
              - $ref: '#/components/schemas/Tasks.SetStep'
              - $ref: '#/components/schemas/Tasks.LogStep'
              - $ref: '#/components/schemas/Tasks.YieldStep'
              - $ref: '#/components/schemas/Tasks.ReturnStep'
              - $ref: '#/components/schemas/Tasks.SleepStep'
              - $ref: '#/components/schemas/Tasks.ErrorWorkflowStep'
              - $ref: '#/components/schemas/Tasks.YieldStep'
              - $ref: '#/components/schemas/Tasks.WaitForInputStep'
              - $ref: '#/components/schemas/Tasks.IfElseWorkflowStep'
              - $ref: '#/components/schemas/Tasks.SwitchStep'
              - $ref: '#/components/schemas/Tasks.ForeachStep'
              - $ref: '#/components/schemas/Tasks.ParallelStep'
              - type: object
                required:
                  - kind_
                  - over
                  - map
                properties:
                  kind_:
                    type: string
                    enum:
                      - map_reduce
                    default: map_reduce
                    readOnly: true
                  over:
                    allOf:
                      - $ref: '#/components/schemas/Common.PyExpression'
                    description: The variable to iterate over
                  map:
                    anyOf:
                      - $ref: '#/components/schemas/Tasks.EvaluateStep'
                      - $ref: '#/components/schemas/Tasks.ToolCallStep'
                      - $ref: '#/components/schemas/Tasks.PromptStep'
                      - $ref: '#/components/schemas/Tasks.GetStep'
                      - $ref: '#/components/schemas/Tasks.SetStep'
                      - $ref: '#/components/schemas/Tasks.LogStep'
                      - $ref: '#/components/schemas/Tasks.YieldStep'
                    description: The steps to run for each iteration
                  reduce:
                    allOf:
                      - $ref: '#/components/schemas/Common.PyExpression'
                    description: |-
                      The expression to reduce the results.
                      If not provided, the results are collected and returned as a list.
                      A special parameter named `results` is the accumulator and `_` is the current value.
                  initial:
                    description: The initial value of the reduce expression
                    default: []
                  parallelism:
                    type: integer
                    format: uint16
                    minimum: 1
                    maximum: 100
                    description: Whether to run the reduce expression in parallel and how many items to run in each batch
                allOf:
                  - type: object
                    required:
                      - kind_
                    properties:
                      kind_:
                        type: string
                        enum:
                          - map_reduce
                        description: The kind of step
                        readOnly: true
                    discriminator:
                      propertyName: kind_
          minItems: 1
          description: The entrypoint of the task.
        input_schema:
          type: object
          additionalProperties: {}
          nullable: true
          description: The schema for the input to the task. `null` means all inputs are valid.
          default: null
        tools:
          type: array
          items:
            $ref: '#/components/schemas/Tasks.TaskTool'
          description: Tools defined specifically for this task not included in the Agent itself.
          default: []
        inherit_tools:
          type: boolean
          description: Whether to inherit tools from the parent agent or not. Defaults to true.
          default: true
        metadata:
          type: object
          additionalProperties: {}
      additionalProperties:
        type: array
        items:
          anyOf:
            - $ref: '#/components/schemas/Tasks.EvaluateStep'
            - $ref: '#/components/schemas/Tasks.ToolCallStep'
            - $ref: '#/components/schemas/Tasks.PromptStep'
            - $ref: '#/components/schemas/Tasks.GetStep'
            - $ref: '#/components/schemas/Tasks.SetStep'
            - $ref: '#/components/schemas/Tasks.LogStep'
            - $ref: '#/components/schemas/Tasks.YieldStep'
            - $ref: '#/components/schemas/Tasks.ReturnStep'
            - $ref: '#/components/schemas/Tasks.SleepStep'
            - $ref: '#/components/schemas/Tasks.ErrorWorkflowStep'
            - $ref: '#/components/schemas/Tasks.YieldStep'
            - $ref: '#/components/schemas/Tasks.WaitForInputStep'
            - $ref: '#/components/schemas/Tasks.IfElseWorkflowStep'
            - $ref: '#/components/schemas/Tasks.SwitchStep'
            - $ref: '#/components/schemas/Tasks.ForeachStep'
            - $ref: '#/components/schemas/Tasks.ParallelStep'
            - type: object
              required:
                - kind_
                - over
                - map
              properties:
                kind_:
                  type: string
                  enum:
                    - map_reduce
                  default: map_reduce
                  readOnly: true
                over:
                  allOf:
                    - $ref: '#/components/schemas/Common.PyExpression'
                  description: The variable to iterate over
                map:
                  anyOf:
                    - $ref: '#/components/schemas/Tasks.EvaluateStep'
                    - $ref: '#/components/schemas/Tasks.ToolCallStep'
                    - $ref: '#/components/schemas/Tasks.PromptStep'
                    - $ref: '#/components/schemas/Tasks.GetStep'
                    - $ref: '#/components/schemas/Tasks.SetStep'
                    - $ref: '#/components/schemas/Tasks.LogStep'
                    - $ref: '#/components/schemas/Tasks.YieldStep'
                  description: The steps to run for each iteration
                reduce:
                  allOf:
                    - $ref: '#/components/schemas/Common.PyExpression'
                  description: |-
                    The expression to reduce the results.
                    If not provided, the results are collected and returned as a list.
                    A special parameter named `results` is the accumulator and `_` is the current value.
                initial:
                  description: The initial value of the reduce expression
                  default: []
                parallelism:
                  type: integer
                  format: uint16
                  minimum: 1
                  maximum: 100
                  description: Whether to run the reduce expression in parallel and how many items to run in each batch
              allOf:
                - type: object
                  required:
                    - kind_
                  properties:
                    kind_:
                      type: string
                      enum:
                        - map_reduce
                      description: The kind of step
                      readOnly: true
                  discriminator:
                    propertyName: kind_
      description: Payload for updating a task
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    ApiKeyAuth_:
      type: apiKey
      in: header
      name: X-Auth-Key
servers:
  - url: https://{serverEnv}.julep.ai/api
    description: The julep cloud service endpoint
    variables:
      serverEnv:
        default: api-alpha
        description: The environment to use
        enum:
          - api
          - api-alpha